Add _load_data method to ParticleDynamicsDataset class
Created a seperate load_data method for the purposes of unit testing the initial loading of the data.
Found out that view reshaped the data differently than the concat method used in colab.
Changed code to be consistent with colab notebook method of reshaping the original data.

Should be changed back to view in the future once training is set up. Want to stick as closely to colab
approach till results are reproduced.