Update ParticleDynamicsDataset to be Initialised Directly from .npy
No longer handles the splitting of the data into train test and validation
sets.

Still expects a root dir with a raw and processed subdir but only outputs
a single processed.pt file in the processed dir.

Design philosophy changed as it made little sense when doing inference as
sometimes it is undesirable to split the input data. Cleaner and more modular
if the dataset just acts to load the raw data without splitting.
Delegate the splitting logic to the user to perform on the raw data. Often
input data can take the form of a large graph which needs to be batched over
/ broken up in a custom way.