Add train.py losses.py and models.py
Modularised the training loop in the new train.py. Loss function defined in losses.py
applies the l1 regularisation, not the l2 weight decay which is applied within the
adam optimiser.

models.py simplified to only contain the gnn for now. HGN to be added.

Basic gnn train config added to repo.