Set N_edge to be Calculated from Message Shape
Personally feel like this makes it more clear and consistent with the
normalisaiton of the MAE with the number of nodes which is the first
dim of the label / pred tensor.

Now we want to normalise the message reg term which we want to normalise
by the first dim of the message tensor.