Add _prune_outlier_graphs to ParticleDynamicsDataset
Added class method which will prune the graph data during the process method call.

This was needed because investigating the NaNs in the KL loss training revealed that
most experiments had simulation data where certain timesteps contained bodies with
velocities or accelerations orders of magnitude greater than the other bodies.
Plotting the distributions of these accelerations and velocities showed huge outliers.

This method removes entire graphs which contain offending bodies.

Had to overwride the _process method of the parent dataset class. This ensures that
changes to this parameter will result in the re processing of the entire dataset during
initialisation in train / test.py.