Update train.py to Stop Logging Loss Params Dict
Currentt logging approach only saves the values calculated over the last batch
for the keys in the loss param dict, want to save average over the entire epoch.
Currently cant think of a neat way to do this and its not particularly important so
just removed that logic right now.