Complete Symbolic Edge Implementation, Update Sparsity plot.
Updated eval_msgs to now save a symbolic_edge.pkl to the output dir.
This dictionary contains keys: {models, important_msg_idxs}. Where models
is a list of pysr regression models and the msg idxs defines which components
of the edge message the models are predicting.

Now also calculate a sanity plot of symbolic  model predictions vs actual edge msgs.
Should basically match if edge messages are correlated with true force but sanity checks
are helpful... Also save results for this in the form of a summary stats json.

Checked that the symbolic edge messages are as close to the nn edge messages as indicated
by these results. Removed the standardisation step in the eval msgs script as it made it
difficult to get the trained pysr model to produce useful outputs in the models message method.
Further, it shouldnt affect the fitting of the symbolic eq since the standardisation is a simple
linear transform (hopefully).

Also changed the sparsity plot to now divide the standard deviations by the sum total of all stds.
This makes it easier to compare what fraction of the total std any given component contributes.